requirements.txt
CacheControl==0.12.6
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
firebase-admin==4.5.3
google-api-core==1.24.1
google-api-python-client==1.12.8
google-auth==1.28.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
google-cloud-core==1.6.0
google-cloud-firestore==2.0.2
google-cloud-storage==1.37.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
grpcio==1.36.1
httplib2==0.18.1
idna==2.10
json5==0.9.5
msgpack==1.0.2
oauthlib==3.1.0
proto-plus==1.18.1
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
six==1.15.0
uritemplate==3.0.1
urllib3==1.26.2
